migrated get file to imaging handling
