+++ NFACT/setup/setup.py
@@ -121,7 +121,7 @@ def return_list_of_subjects_from_file(path_to_list: str) -> list:
         if path_to_list.split(".")[1] != "txt":
             error_and_exit(
                 False,
-                "List of subjects is not ascii file. Please specify a list of subject.",
+                "List of subjects is not a text file. Please specify a list of subject.",
             )
     # Hacky way to allow sub list not to have an extension
     except IndexError:
@@ -131,3 +131,10 @@ def return_list_of_subjects_from_file(path_to_list: str) -> list:
     except Exception as e:
         error_and_exit(False, f"Unable to open subject list due to: {e}")
     return list_of_subjects
+
+
+def process_seeds(seeds):
+    try:
+        seeds = read_file_to_list(seeds)
+    except Exception as e:
+        error_and_exit(False, f"Unable to read seeds text file due to {e}")
